services/tts-service/requirements.txt
fastapi
uvicorn
boto3
gtts
kafka-python
uuid
